FISH-10919 Rework Docker runner
The Platform TCK is run differently to how Faces is run.
The Faces TCK comes with its own runner pom which we replace/update with our own and then run. The Platform TCK does not have this - you must provide your own runner from scratch.
This poses an issue in that our poms are not standalone - they are expected to run as a part of a full project
To get around this we provide the full maven project to Docker - we essentially clone the current state of the runner repository in Docker.

This commit also cleans up some copy-paste errors.